Crate zerotrie

A data structure offering zero-copy storage and retrieval of byte strings, with a focus on the efficient storage of ASCII strings. Strings are mapped to usize values.

ZeroTrie does not support mutation because doing so would require recomputing the entire data structure. Instead, it supports conversion to and from LiteMap and BTreeMap.

There are multiple variants of ZeroTrie optimized for different use cases.

Examples

use zerotrie::ZeroTrie;

let data: &[(&str, usize)] = &[("abc", 11), ("xyz", 22), ("axyb", 33)];

let trie: ZeroTrie<Vec<u8>> = data.iter().copied().collect();

assert_eq!(trie.get("axyb"), Some(33));
assert_eq!(trie.byte_len(), 18);

Internal Structure

To read about the internal structure of ZeroTrie, build the docs with private modules:

cargo doc --document-private-items --all-features --no-deps --open

Modules

cursor

Types for walking stepwise through a trie.

Structs

ZeroAsciiIgnoreCaseTrie

A data structure that compactly maps from ASCII strings to integers in a case-insensitive way.

ZeroTrie

A data structure that compactly maps from byte sequences to integers.

ZeroTrieExtendedCapacity

A data structure that maps from a large number of byte strings to integers.

ZeroTriePerfectHash

A data structure that compactly maps from byte strings to integers.

ZeroTrieSimpleAscii

A data structure that compactly maps from ASCII strings to integers.

Enums

ZeroTrieBuildError

Error types for the zerotrie crate.