zerotrie::builder::konst::builder

Struct ZeroTrieBuilderConst

pub(crate) struct ZeroTrieBuilderConst<const N: usize> {
    data: ConstArrayBuilder<N, u8>,
}

A low-level builder for [ZeroTrieSimpleAscii]. Works in const contexts.

All methods that grow the trie will panic if the capacity N is not enough.

Fields
data: ConstArrayBuilder<N, u8>
Implementations

impl<const N: usize> ZeroTrieBuilderConst<N>

pub const fn build_or_panic(self) -> [u8; N]

Returns the trie data, panicking if the buffer is the wrong size.

pub const fn new() -> Self

Creates a new empty builder.

pub const fn prepend_ascii(&mut self, ascii: u8) -> usize

Prepends an ASCII node to the front of the builder. Returns the new builder and the delta in length, which is always 1.

pub const fn prepend_value(&mut self, value: usize) -> usize

Prepends a value node to the front of the builder. Returns the new builder and the delta in length, which depends on the size of the varint.

pub const fn prepend_branch(&mut self, value: usize) -> usize

Prepends a branch node to the front of the builder. Returns the new builder and the delta in length, which depends on the size of the varint.

pub const fn prepend_slice(&mut self, s: ConstSlice<'_, u8>) -> usize

Prepends multiple arbitrary bytes to the front of the builder. Returns the new builder and the delta in length, which is the length of the slice.

pub const fn prepend_n_zeros(&mut self, n: usize)

Prepends multiple zeros to the front of the builder. Returns the new builder.

pub const fn bitor_assign_at_or_panic(&mut self, index: usize, bits: u8)

Performs the operation self[index] |= bits

pub const fn from_tuple_slice<'a, const K: usize>( items: ByteSliceWithIndices<'a>, ) -> Self

Creates a new builder containing the elements in the given slice of key/value pairs.

K is the stack size of the lengths stack. If you get an error such as AsciiTrie Builder: Need more stack, try increasing K.

Panics

Panics if the items are not sorted

pub const fn from_slice_with_indices<const K: usize>( items: ByteSliceWithIndices<'_>, ) -> Self

Creates a new builder containing the elements in the given slice of key/value pairs.

Assumes that the items are sorted. If they are not, unexpected behavior may occur.

K is the stack size of the lengths stack. If you get an error such as AsciiTrie Builder: Need more stack, try increasing K.

const fn create_or_panic<const K: usize>( &mut self, all_items: ByteSliceWithIndices<'_>, ) -> usize

The actual builder algorithm. For an explanation, see crate::builder.