Macro zerotrie::builder::konst::store::const_for_each

macro_rules! const_for_each {
    ($safe_const_slice:expr, $item:tt, $inner:expr) => { ... };
}

Evaluates a block over each element of a const slice. Takes three arguments:

Expression that resolves to the ConstSlice.

Token that will be assigned the value of the element.

Block to evaluate for each element.