Struct zerotrie::builder::nonconst::builder::ZeroTrieBuilder

pub(crate) struct ZeroTrieBuilder<S> {
    data: S,
    phf_cache: PerfectByteHashMapCacheOwned,
    options: ZeroTrieBuilderOptions,
}

A low-level builder for ZeroTrie. Supports all options.

Fields
data: Sphf_cache: PerfectByteHashMapCacheOwnedoptions: ZeroTrieBuilderOptions
Implementations

impl<S: TrieBuilderStore> ZeroTrieBuilder<S>

pub fn to_bytes(&self) -> Vec<u8>

Returns the trie data as a Vec<u8>.

fn prepend_ascii(&mut self, ascii: u8) -> Result<usize, Error>

Prepends a byte value to the front of the builder. If it is ASCII, an ASCII node is prepended. If it is non-ASCII, if there is already a span node at the front, we modify the span node to add the new byte; otherwise, we create a new span node. Returns the delta in length, which is either 1 or 2.

fn prepend_value(&mut self, value: usize) -> usize

Prepends a value node to the front of the builder. Returns the delta in length, which depends on the size of the varint.

fn prepend_branch(&mut self, value: usize) -> usize

Prepends a branch node to the front of the builder. Returns the delta in length, which depends on the size of the varint.

fn prepend_slice(&mut self, s: &[u8]) -> usize

Prepends multiple arbitrary bytes to the front of the builder. Returns the delta in length, which is the length of the slice.

pub fn from_bytes_iter<K: AsRef<[u8]>, I: IntoIterator<Item = (K, usize)>>( iter: I, options: ZeroTrieBuilderOptions, ) -> Result<Self, Error>

Builds a ZeroTrie from an iterator of bytes. It first collects and sorts the iterator.

pub fn from_sorted_tuple_slice( items: &[(&ByteStr, usize)], options: ZeroTrieBuilderOptions, ) -> Result<Self, Error>

Builds a ZeroTrie with the given items and options. Assumes that the items are sorted, except for a case-insensitive trie where the items are re-sorted.

Panics

May panic if the items are not sorted.

fn from_sorted_tuple_slice_impl( items: &[(&ByteStr, usize)], options: ZeroTrieBuilderOptions, ) -> Result<Self, Error>

Internal constructor that does not re-sort the items.

fn create(&mut self, all_items: &[(&ByteStr, usize)]) -> Result<usize, Error>

The actual builder algorithm. For an explanation, see crate::builder.