zerotrie::cursor

Struct ZeroAsciiIgnoreCaseTrieCursor

pub struct ZeroAsciiIgnoreCaseTrieCursor<'a> { /* private fields */ }

A cursor into a ZeroAsciiIgnoreCaseTrie, useful for stepwise lookup.

For examples, see ZeroAsciiIgnoreCaseTrie::cursor().

Implementations

impl ZeroAsciiIgnoreCaseTrieCursor<'_>

pub fn step(&mut self, byte: u8) -> Option<u8>

Steps the cursor one byte into the trie.

Returns the byte if matched, which may be a different case than the input byte. If this function returns None, any lookup loops can be terminated.

Examples

Normalize the case of a value by stepping through an ignore-case trie:

use std::borrow::Cow;
use zerotrie::ZeroAsciiIgnoreCaseTrie;

// A trie with two values: "aBc" and "aBcdEf"
let trie = ZeroAsciiIgnoreCaseTrie::from_bytes(b"aBc\x80dEf\x81");

// Get out the value for "abc" and normalize the key string
let mut cursor = trie.cursor();
let mut key_str = Cow::Borrowed("abc".as_bytes());
let mut i = 0;
let value = loop {
    let Some(&input_byte) = key_str.get(i) else {
        break cursor.take_value();
    };
    let Some(matched_byte) = cursor.step(input_byte) else {
        break None;
    };
    if matched_byte != input_byte {
        key_str.to_mut()[i] = matched_byte;
    }
    i += 1;
};

assert_eq!(value, Some(0));
assert_eq!(&*key_str, "aBc".as_bytes());

For more examples, see ZeroTrieSimpleAsciiCursor::step.

pub fn take_value(&mut self) -> Option<usize>

Takes the value at the current position.

For more details, see ZeroTrieSimpleAsciiCursor::take_value.

pub fn probe(&mut self, index: usize) -> Option<AsciiProbeResult>

Probes the next byte in the cursor.

For more details, see ZeroTrieSimpleAsciiCursor::probe.

pub fn is_empty(&self) -> bool

Checks whether the cursor points to an empty trie.

For more details, see ZeroTrieSimpleAsciiCursor::is_empty.