zerotrie

Module options

Options for building and reading from a ZeroTrie.

These options are internal to the crate. A small selection of options are exported by way of the different public types on this crate.

Structs

ZeroTrieBuilderOptions 🔒

Enums

AsciiMode 🔒

Whether to support non-ASCII data in the ZeroTrie.

CapacityMode 🔒

Whether to enforce a limit to the capacity of the ZeroTrie.

CaseSensitivity 🔒

How to handle strings with mixed ASCII case at a node, such as “abc” and “Abc”

PhfMode 🔒

Whether to use the perfect hash function in the ZeroTrie.

Traits

ZeroTrieWithOptions 🔒