zerotrie::reader

Function get_branch

fn get_branch(trie: &[u8], i: usize, n: usize, w: usize) -> &[u8] 

Given a slice starting with an offset table, returns the trie for the given index.

Arguments:

trie = a trie pointing at an offset table (after the branch node and search table)

i = the desired index within the offset table

n = the number of items in the offset table

w = the width of the offset table items minus one