Struct zerotrie::ZeroAsciiIgnoreCaseTrie

#[repr(transparent)]
pub struct ZeroAsciiIgnoreCaseTrie<Store: ?Sized> { /* private fields */ }

A data structure that compactly maps from ASCII strings to integers in a case-insensitive way.

Examples

use litemap::LiteMap;
use zerotrie::ZeroAsciiIgnoreCaseTrie;

let mut map = LiteMap::new_vec();
map.insert(&b"foo"[..], 1);
map.insert(b"Bar", 2);
map.insert(b"Bazzoo", 3);

let trie = ZeroAsciiIgnoreCaseTrie::try_from(&map)?;

assert_eq!(trie.get(b"foo"), Some(1));
assert_eq!(trie.get(b"bar"), Some(2));
assert_eq!(trie.get(b"BAR"), Some(2));
assert_eq!(trie.get(b"bazzoo"), Some(3));
assert_eq!(trie.get(b"unknown"), None);

Strings with different cases of the same character at the same offset are not allowed:

use litemap::LiteMap;
use zerotrie::ZeroAsciiIgnoreCaseTrie;

let mut map = LiteMap::new_vec();
map.insert(&b"bar"[..], 1);
// OK: 'r' and 'Z' are different letters
map.insert(b"baZ", 2);
// Bad: we already inserted 'r' so we cannot also insert 'R' at the same position
map.insert(b"baR", 2);

ZeroAsciiIgnoreCaseTrie::try_from(&map).expect_err("mixed-case strings!");

Implementations

impl<Store> ZeroAsciiIgnoreCaseTrie<Store>
where Store: AsRef<[u8]> + ?Sized,

pub fn cursor(&self) -> ZeroAsciiIgnoreCaseTrieCursor<'_>

Gets a cursor into the current trie.

Useful to query a trie with data that is not a slice.

This is currently supported only on ZeroTrieSimpleAscii and ZeroAsciiIgnoreCaseTrie.

Examples

Get a value out of a trie by writing it to the cursor:

use core::fmt::Write;
use zerotrie::ZeroAsciiIgnoreCaseTrie;

// A trie with two values: "aBc" and "aBcdEf"
let trie = ZeroAsciiIgnoreCaseTrie::from_bytes(b"aBc\x80dEf\x81");

// Get out the value for "abc" (case-insensitive!)
let mut cursor = trie.cursor();
write!(&mut cursor, "abc");
assert_eq!(cursor.take_value(), Some(0));

For more examples, see ZeroTrieSimpleAscii::cursor.

impl<'a> ZeroAsciiIgnoreCaseTrie<&'a [u8]>

pub fn into_cursor(self) -> ZeroAsciiIgnoreCaseTrieCursor<'a>

Same as ZeroAsciiIgnoreCaseTrie::cursor() but moves self to avoid having to doubly anchor the trie to the stack.

impl<Store> ZeroAsciiIgnoreCaseTrie<Store>

pub const fn from_store(store: Store) -> Self

Create a trie directly from a store.

If the store does not contain valid bytes, unexpected behavior may occur.

pub fn take_store(self) -> Store

Takes the byte store from this trie.

pub fn convert_store<X: From<Store>>(self) -> ZeroAsciiIgnoreCaseTrie<X>

Converts this trie’s store to a different store implementing the From trait.

For example, use this to change ZeroAsciiIgnoreCaseTrie<Vec<u8>> to ZeroAsciiIgnoreCaseTrie<Cow<[u8]>>.

Examples

use std::borrow::Cow;
use zerotrie::ZeroAsciiIgnoreCaseTrie;

let trie: ZeroAsciiIgnoreCaseTrie<Vec<u8>> = ZeroAsciiIgnoreCaseTrie::from_bytes(b"abc\x85").to_owned();
let cow: ZeroAsciiIgnoreCaseTrie<Cow<[u8]>> = trie.convert_store();

assert_eq!(cow.get(b"abc"), Some(5));

impl<Store> ZeroAsciiIgnoreCaseTrie<Store>
where Store: AsRef<[u8]> + ?Sized,

pub fn get<K>(&self, key: K) -> Option<usize>
where K: AsRef<[u8]>,

Queries the trie for a string.

pub fn is_empty(&self) -> bool

Returns true if the trie is empty.

pub fn byte_len(&self) -> usize

Returns the size of the trie in number of bytes.

To get the number of keys in the trie, use .iter().count():

use zerotrie::ZeroAsciiIgnoreCaseTrie;

// A trie with two values: "abc" and "abcdef"
let trie: &ZeroAsciiIgnoreCaseTrie<[u8]> = ZeroAsciiIgnoreCaseTrie::from_bytes(b"abc\x80def\x81");

assert_eq!(8, trie.byte_len());
assert_eq!(2, trie.iter().count());

pub fn as_bytes(&self) -> &[u8]

Returns the bytes contained in the underlying store.

pub fn as_borrowed(&self) -> &ZeroAsciiIgnoreCaseTrie<[u8]>

Returns this trie as a reference transparent over a byte slice.

pub fn as_borrowed_slice(&self) -> ZeroAsciiIgnoreCaseTrie<&[u8]>

Returns a trie with a store borrowing from this trie.

impl<Store> ZeroAsciiIgnoreCaseTrie<Store>
where Store: AsRef<[u8]> + ?Sized,

pub fn to_owned(&self) -> ZeroAsciiIgnoreCaseTrie<Vec<u8>>

Converts a possibly-borrowed $name to an owned one.

✨ Enabled with the alloc Cargo feature.

Examples

use zerotrie::ZeroAsciiIgnoreCaseTrie;

let trie: &ZeroAsciiIgnoreCaseTrie<[u8]> = ZeroAsciiIgnoreCaseTrie::from_bytes(b"abc\x85");
let owned: ZeroAsciiIgnoreCaseTrie<Vec<u8>> = trie.to_owned();

assert_eq!(trie.get(b"abc"), Some(5));
assert_eq!(owned.get(b"abc"), Some(5));

pub fn iter(&self) -> impl Iterator<Item = (String, usize)> + '_

Returns an iterator over the key/value pairs in this trie.

✨ Enabled with the alloc Cargo feature.

Examples

use zerotrie::ZeroAsciiIgnoreCaseTrie;

// A trie with two values: "abc" and "abcdef"
let trie: &ZeroAsciiIgnoreCaseTrie<[u8]> = ZeroAsciiIgnoreCaseTrie::from_bytes(b"abc\x80def\x81");

let mut it = trie.iter();
assert_eq!(it.next(), Some(("abc".into(), 0)));
assert_eq!(it.next(), Some(("abcdef".into(), 1)));
assert_eq!(it.next(), None);

impl ZeroAsciiIgnoreCaseTrie<[u8]>

pub fn from_bytes(trie: &[u8]) -> &Self

Casts from a byte slice to a reference to a trie with the same lifetime.

If the bytes are not a valid trie, unexpected behavior may occur.

impl<Store> ZeroAsciiIgnoreCaseTrie<Store>
where Store: AsRef<[u8]> + ?Sized,

pub fn to_btreemap(&self) -> BTreeMap<String, usize>

Exports the data from this ZeroTrie type into a BTreeMap.

✨ Enabled with the alloc Cargo feature.

Examples

use zerotrie::ZeroAsciiIgnoreCaseTrie;
use std::collections::BTreeMap;

let trie = ZeroAsciiIgnoreCaseTrie::from_bytes(b"abc\x81def\x82");
let items = trie.to_btreemap();

assert_eq!(items.len(), 2);

let recovered_trie: ZeroAsciiIgnoreCaseTrie<Vec<u8>> = items
    .into_iter()
    .collect();
assert_eq!(trie.as_bytes(), recovered_trie.as_bytes());

impl<Store> ZeroAsciiIgnoreCaseTrie<Store>
where Store: AsRef<[u8]> + ?Sized,

pub fn to_litemap(&self) -> LiteMap<String, usize>

Exports the data from this ZeroTrie type into a LiteMap.

✨ Enabled with the litemap Cargo feature.

Examples

use zerotrie::ZeroAsciiIgnoreCaseTrie;
use litemap::LiteMap;

let trie = ZeroAsciiIgnoreCaseTrie::from_bytes(b"abc\x81def\x82");

let items = trie.to_litemap();
assert_eq!(items.len(), 2);

let recovered_trie: ZeroAsciiIgnoreCaseTrie<Vec<u8>> = items
    .iter()
    .map(|(k, v)| (k, *v))
    .collect();
assert_eq!(trie.as_bytes(), recovered_trie.as_bytes());