Struct zerotrie::ZeroTrie

pub struct ZeroTrie<Store>(/* private fields */);

A data structure that compactly maps from byte sequences to integers.

There are several variants of ZeroTrie which are very similar but are optimized for different use cases:

ZeroTrieSimpleAscii is the most compact structure. Very fast for small data. Only stores ASCII-encoded strings. Can be const-constructed!

ZeroTriePerfectHash is also compact, but it also supports arbitrary binary strings. It also scales better to large data. Cannot be const-constructed.

ZeroTrieExtendedCapacity can be used if more than 2^32 bytes are required.

You can create a ZeroTrie directly, in which case the most appropriate backing implementation will be chosen.

Backing Store

The data structure has a flexible backing data store. The only requirement for most functionality is that it implement AsRef<[u8]>. All of the following are valid ZeroTrie types:

ZeroTrie<[u8]> (dynamically sized type: must be stored in a reference or Box)

ZeroTrie<&[u8]> (borrows its data from a u8 buffer)

ZeroTrie<Vec<u8>> (fully owned data)

ZeroTrie<ZeroVec<u8>> (the recommended borrowed-or-owned signature)

Cow<ZeroTrie<[u8]>> (another borrowed-or-owned signature)

ZeroTrie<Cow<[u8]>> (another borrowed-or-owned signature)

Examples

use litemap::LiteMap;
use zerotrie::ZeroTrie;

let mut map = LiteMap::<&[u8], usize>::new_vec();
map.insert("foo".as_bytes(), 1);
map.insert("bar".as_bytes(), 2);
map.insert("bazzoo".as_bytes(), 3);

let trie = ZeroTrie::try_from(&map)?;

assert_eq!(trie.get("foo"), Some(1));
assert_eq!(trie.get("bar"), Some(2));
assert_eq!(trie.get("bazzoo"), Some(3));
assert_eq!(trie.get("unknown"), None);

Implementations

impl<Store> ZeroTrie<Store>

pub fn take_store(self) -> Store

Takes the byte store from this trie.

pub fn convert_store<NewStore>(self) -> ZeroTrie<NewStore>
where NewStore: From<Store>,

Converts this trie’s store to a different store implementing the From trait.

For example, use this to change ZeroTrie<Vec<u8>> to ZeroTrie<Cow<[u8]>>.

impl<Store> ZeroTrie<Store>
where Store: AsRef<[u8]>,

pub fn get<K>(&self, key: K) -> Option<usize>
where K: AsRef<[u8]>,

Queries the trie for a string.

pub fn is_empty(&self) -> bool

Returns true if the trie is empty.

pub fn byte_len(&self) -> usize

Returns the size of the trie in number of bytes.

To get the number of keys in the trie, use .iter().count().

impl<Store> ZeroTrie<Store>
where Store: AsRef<[u8]>,

pub fn to_btreemap(&self) -> BTreeMap<Box<[u8]>, usize>

Exports the data from this ZeroTrie into a BTreeMap.

impl<Store> ZeroTrie<Store>
where Store: AsRef<[u8]>,

pub fn to_litemap(&self) -> LiteMap<Box<[u8]>, usize>

Exports the data from this ZeroTrie into a LiteMap.