zerotrie

Struct ZeroTrieExtendedCapacity

pub struct ZeroTrieExtendedCapacity<Store: ?Sized> { /* private fields */ }

A data structure that maps from a large number of byte strings to integers.

For more information, see ZeroTrie.

Implementations

impl<Store> ZeroTrieExtendedCapacity<Store>

pub const fn into_zerotrie(self) -> ZeroTrie<Store>

Wrap this specific ZeroTrie variant into a ZeroTrie.

impl<Store> ZeroTrieExtendedCapacity<Store>

pub const fn from_store(store: Store) -> Self

Create a trie directly from a store.

If the store does not contain valid bytes, unexpected behavior may occur.

pub fn into_store(self) -> Store

Takes the byte store from this trie.

pub fn convert_store<X: From<Store>>(self) -> ZeroTrieExtendedCapacity<X>

Converts this trieโs store to a different store implementing the From trait.

For example, use this to change ZeroTrieExtendedCapacity<Vec<u8>> to ZeroTrieExtendedCapacity<Cow<[u8]>>.

Examples

use std::borrow::Cow;
use zerotrie::ZeroTrieExtendedCapacity;

let trie: ZeroTrieExtendedCapacity<Vec<u8>> = ZeroTrieExtendedCapacity::from_bytes(b"abc\x85").to_owned();
let cow: ZeroTrieExtendedCapacity<Cow<[u8]>> = trie.convert_store();

assert_eq!(cow.get(b"abc"), Some(5));

impl<Store> ZeroTrieExtendedCapacity<Store>
where Store: AsRef<[u8]> + ?Sized,

pub fn get<K>(&self, key: K) -> Option<usize>
where K: AsRef<[u8]>,

Queries the trie for a string.

pub fn is_empty(&self) -> bool

Returns true if the trie is empty.

pub fn byte_len(&self) -> usize

Returns the size of the trie in number of bytes.

To get the number of keys in the trie, use .iter().count():

use zerotrie::ZeroTrieExtendedCapacity;

// A trie with two values: "abc" and "abcdef"
let trie: &ZeroTrieExtendedCapacity<[u8]> = ZeroTrieExtendedCapacity::from_bytes(b"abc\x80def\x81");

assert_eq!(8, trie.byte_len());
assert_eq!(2, trie.iter().count());

pub fn as_bytes(&self) -> &[u8]

Returns the bytes contained in the underlying store.

pub fn as_borrowed(&self) -> &ZeroTrieExtendedCapacity<[u8]>

Returns this trie as a reference transparent over a byte slice.

pub fn as_borrowed_slice(&self) -> ZeroTrieExtendedCapacity<&[u8]>

Returns a trie with a store borrowing from this trie.

impl ZeroTrieExtendedCapacity<[u8]>

pub fn from_bytes(trie: &[u8]) -> &Self

Casts from a byte slice to a reference to a trie with the same lifetime.

If the bytes are not a valid trie, unexpected behavior may occur.