Struct zerotrie::builder::konst::builder::ZeroTrieBuilderConst

pub(crate) struct ZeroTrieBuilderConst<const N: usize> {
    data: ConstArrayBuilder<N, u8>,
}

A low-level builder for ZeroTrieSimpleAscii. Works in const contexts.

Fields
data: ConstArrayBuilder<N, u8>
Implementations

impl<const N: usize> ZeroTrieBuilderConst<N>

pub const fn take_or_panic(self) -> [u8; N]

Returns the trie data, panicking if the buffer is the wrong size.

pub const fn new() -> Self

Creates a new empty builder.

const fn prepend_ascii(self, ascii: u8) -> (Self, usize)

Prepends an ASCII node to the front of the builder. Returns the new builder and the delta in length, which is always 1.

const fn prepend_value(self, value: usize) -> (Self, usize)

Prepends a value node to the front of the builder. Returns the new builder and the delta in length, which depends on the size of the varint.

const fn prepend_branch(self, value: usize) -> (Self, usize)

Prepends a branch node to the front of the builder. Returns the new builder and the delta in length, which depends on the size of the varint.

const fn prepend_slice(self, s: ConstSlice<'_, u8>) -> (Self, usize)

Prepends multiple arbitrary bytes to the front of the builder. Returns the new builder and the delta in length, which is the length of the slice.

const fn prepend_n_zeros(self, n: usize) -> Self

Prepends multiple zeros to the front of the builder. Returns the new builder.

const fn bitor_assign_at(self, index: usize, bits: u8) -> Self

Performs the operation self[index] |= bits

pub const fn from_tuple_slice<'a, const K: usize>( items: &[(&'a ByteStr, usize)], ) -> Result<Self, Error>

Creates a new builder containing the elements in the given slice of key/value pairs.

K is the stack size of the lengths stack. If you get an error such as “AsciiTrie Builder: Need more stack”, try increasing K.

Panics

Panics if the items are not sorted

pub const fn from_sorted_const_tuple_slice<const K: usize>( items: ConstSlice<'_, (&ByteStr, usize)>, ) -> Result<Self, Error>

Creates a new builder containing the elements in the given slice of key/value pairs.

Assumes that the items are sorted. If they are not, unexpected behavior may occur.

K is the stack size of the lengths stack. If you get an error such as “AsciiTrie Builder: Need more stack”, try increasing K.

const fn create_or_panic<const K: usize>( self, all_items: ConstSlice<'_, (&ByteStr, usize)>, ) -> (Self, usize)

The actual builder algorithm. For an explanation, see crate::builder.