Module zerotrie::zerotrie

Macros

impl_dispatch 🔒

impl_zerotrie_subtype 🔒

Structs

ZeroAsciiIgnoreCaseTrie

A data structure that compactly maps from ASCII strings to integers in a case-insensitive way.

ZeroTrie

A data structure that compactly maps from byte sequences to integers.

ZeroTrieExtendedCapacity

A data structure that maps from a large number of byte strings to integers.

ZeroTriePerfectHash

A data structure that compactly maps from byte strings to integers.

ZeroTrieSimpleAscii

A data structure that compactly maps from ASCII strings to integers.

Enums

ZeroTrieFlavor 🔒

Functions

string_to_box_u8 🔒